Check, if the dimension allows mining of a block
This was an issue with the Dimensional Dungeons mod, which creates a
separate dimension, where most blocks are unbreakable. It achieves this
by overriding the canMineBlock function of the Dimension class.

Now MiningGadgets also runs this function to check, if the block should be mineable.
Furthermore the RenderBlock is now excluded from the checks, as it would be blocked
by the Dimensional Dungeons mod, as it is not contained in the block whitelist. If it wouldn't
be excluded, it wouldn't be considered to be broken, which prevents the contained block
from being broken.